projects/transformer-keyword-labeling/requirements.txt
torch>=2.0
transformers
seqeval
scikit-learn
tqdm
pyyaml
